requirements.txt
asgiref==3.3.4
Django==3.2.4
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-debug-toolbar-force==0.1.8
django-environ==0.4.5
django-filter==2.4.0
django-js-asset==1.2.2
django-mptt==0.12.0
django-nine==0.2.4
django-redis==5.0.0
djangorestframework==3.12.4
djangorestframework-recursive==0.1.2
gunicorn==20.1.0
importlib-metadata==4.5.0
Markdown==3.3.4
mysqlclient==2.0.3
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pytz==2021.1
redis==3.5.3
rest-framework-generic-relations==2.0.0
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
zipp==3.4.1
transliterate==1.10.2